renames and cleanups
