+++ src/server/js/resources/resource.js
