+++ src/server/js/server.js
@@ -46,5 +46,3 @@ configure(app);
 app.listen(PORT, () => {
     console.log(`Listening on port ${PORT}`);
 });
-
-import * as r from './resources/houses/resolvers';
